[Sunny] Step4

+++ src/main/java/com/codessquad/qna/question/QuestionController.java
@@ -1,21 +1,25 @@
 package com.codessquad.qna.question;
 
-import com.codessquad.qna.sessionutils.HttpSessionUtils;
+import com.codessquad.qna.answer.AnswerRepository;
+import com.codessquad.qna.utils.HttpSessionUtils;
 import com.codessquad.qna.user.User;
+import com.codessquad.qna.utils.Result;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpSession;
 
 @Controller
 @RequestMapping("/questions")
 public class QuestionController {
-    private static Logger log = LoggerFactory.getLogger(QuestionController.class);
+    private Logger log = LoggerFactory.getLogger(QuestionController.class);
+
+    @Autowired
+    private AnswerRepository answerRepository;
 
     @Autowired
     private QuestionRepository questionRepository;
@@ -30,40 +34,77 @@ public class QuestionController {
     }
 
     @PostMapping("")
-    public String createQna(String title, String contents, HttpSession session) {
+    public String createQuestion(String title, String contents, HttpSession session) {
         if (HttpSessionUtils.isNoneExistentUser(session)) {
             return "redirect:/users/loginForm";
         }
 
         User sessionUser = HttpSessionUtils.getUserFromSession(session);
-        Question newQuestion = new Question(sessionUser.getUserId(), title, contents);
+        Question newQuestion = new Question(sessionUser, title, contents);
         questionRepository.save(newQuestion);
         return "redirect:/";
     }
 
     @GetMapping("/{id}")
-    public String showQuestionContents(@PathVariable Long id, Model model) throws IllegalAccessException {
-        model.addAttribute("question", questionRepository.findById(id).orElseThrow(IllegalAccessException::new));
+    public String showQuestionContents(@PathVariable Long id, Model model) {
+        model.addAttribute("question",
+                questionRepository.findById(id).orElseThrow(IllegalStateException::new));
+
         return "qna/show";
     }
 
     @GetMapping("/{id}/form")
-    public String updateForm(@PathVariable Long id, Model model) throws IllegalAccessException {
-        model.addAttribute("question", questionRepository.findById(id).orElseThrow(IllegalAccessException::new));
-        return "/qna/updateForm";
+    public String updateForm(@PathVariable Long id, Model model, HttpSession session) {
+        Question question = questionRepository.findById(id).orElseThrow(IllegalStateException::new);
+        Result result = valid(session, question);
+        if (result.isNotValid()) {
+            model.addAttribute("errorMessage", result.getErrorMessage());
+            return "user/login";
+        }
+
+        model.addAttribute("question", question);
+        return "qna/updateForm";
     }
 
     @PutMapping("/{id}")
-    public String updateQna(@PathVariable Long id, String title, String contents) throws IllegalAccessException {
-        Question question = questionRepository.findById(id).orElseThrow(IllegalAccessException::new);
+    public String updateQuestion(@PathVariable Long id, String title, String contents,
+                                 Model model, HttpSession session) {
+        Question question = questionRepository.findById(id).orElseThrow(IllegalStateException::new);
+        Result result = valid(session, question);
+        if (result.isNotValid()) {
+            model.addAttribute("errorMessage", result.getErrorMessage());
+            return "user/login";
+        }
+
+        model.addAttribute("question", question);
         question.update(title, contents);
         questionRepository.save(question);
-        return String.format("redirect:/questions/%d",id);
+        return "redirect:/questions/" + id;
     }
 
     @DeleteMapping("/{id}")
-    public String deleteQna(@PathVariable Long id) {
+    public String deleteQuestion(@PathVariable Long id, Model model, HttpSession session) {
+        Question question = questionRepository.findById(id).orElseThrow(IllegalStateException::new);
+        Result result = valid(session, question);
+        if (result.isNotValid()) {
+            model.addAttribute("errorMessage", result.getErrorMessage());
+            return "user/login";
+        }
+
         questionRepository.deleteById(id);
         return "redirect:/";
     }
+
+    private Result valid(HttpSession session, Question question) {
+        if (HttpSessionUtils.isNoneExistentUser(session)) {
+            return Result.fail("로그인이 필요합니다.");
+        }
+
+        User loginUser = HttpSessionUtils.getUserFromSession(session);
+        if (question.isNotSameWriter(loginUser)) {
+            return Result.fail("자신의 쓴 글만 수정, 삭제가 가능합니다.");
+        }
+        return Result.ok();
+    }
+
 }
